// CodeLabFrontend/src/app/shared/enums/mensagem.enum.ts
export enum EMensagem {
  ALGO_DEU_ERRADO = 'Oops...! Algo deu errado.',
  ALTERACOES_PENDENTES_TITULO = 'Alterações Pendentes',
  ALTERACOES_PENDENTES_TEXTO = 'Você possui alterações que não foram salvas, deseja mesmo sair da página?',
  CAMPOS_LOGIN_INVALIDOS = 'Email e senha devem ser preenchidos corretamente!',
  CAMPOS_NAO_PREENCHIDOS = 'Campos obrigatórios não preenchidos!',
  FECHAR = 'fechar',
  REGISTRO_NAO_ENCONTRADO = 'Registro não encontrado!',
  SESSAO_EXPIRADA = 'Sessão expirada!',
  ADICIONAR_VENDA_ITEM = 'Adicione ao menos um item à venda!',
  SENHA_ALTERADA_SUCESSO = 'Senha alterada com sucesso!',
}